% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_redes_de_palabras.R
\name{crear_redes_de_palabras}
\alias{crear_redes_de_palabras}
\title{Crea redes usando algún vector de palabras u oraciones para ser analizadas}
\usage{
crear_redes_de_palabras(x)
}
\arguments{
\item{x}{Un vector de carácteres.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Utilizando la librería \code{tm} y \code{igraph}, opta por un 'wrapper' para generar
redes listas para ser utilizadas bajo el contexto de procesamiento de lenguaje
natural.
}
