% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_tokenizer.R
\name{fun_tokenizer}
\alias{fun_tokenizer}
\title{Función interna tokenizadora para la extracción de palabras}
\usage{
fun_tokenizer(j, n_words)
}
\arguments{
\item{j}{Vector de carácteres.}

\item{n_words}{Número de palabras conjuntas.}
}
\description{
Esta función es usada de manera interna por \code{obtener_terminos()}
}
