% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_terminos.R
\name{get_terms}
\alias{get_terms}
\title{Obtener frecuencia de palabras}
\usage{
get_terms(i, n_words = 1)
}
\arguments{
\item{i}{vector de carácteres.}

\item{n_words}{Número de palabras combinadas.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
