% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_red.R
\name{grafico_red_hubs}
\alias{grafico_red_hubs}
\title{Grafica una red usando score de Hubs}
\usage{
grafico_red_hubs(red)
}
\arguments{
\item{red}{Un objeto de clase \code{igraph} resultante de la función \code{crear_redes_de_palabras()} o
de procesos similares usando el paquete \code{igraph}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Utilizando un objeto resultante de la función \code{crear_redes_de_palabras()},
crea un gráfico de redes usando el score de Hubs. Para esto,
se usan bajo la capa la librería \code{igraph}.
}
