% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labinstrumentos-package.R
\docType{package}
\name{labinstrumentos-package}
\alias{labinstrumentos}
\alias{labinstrumentos-package}
\title{Laboratorio de Instrumentos}
\description{
El objetivo de este paquete es favoreces el libre acceso los datos y demás herramientas asistenciales
para el análisis y así, generar una propuesta de recomendaciones a los instrumentos ANID más utilizados
en la macrozona teniendo en cuenta las brechas regionales y el sistema CTCI local.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SEREMICTCI/labinstrumentos}
  \item Report bugs at \url{https://github.com/SEREMICTCI/labinstrumentos/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Carlos Morales Quiroz \email{csmorales@minciencia.gob.cl} [contributor]
}

}
\keyword{internal}
