% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labinstrumentos.R
\docType{data}
\name{labinstrumentos}
\alias{labinstrumentos}
\title{Datos del Laboratorio de Instrumentos}
\format{
Un data frame con 158 filas y 11 variables:
\itemize{
\item \code{laboratorio}: ...
\item \code{grupo}: ...
\item \code{problema}: ...
\item \code{cluster_manual_problema}: ...
\item \code{clean_problema}: ...
\item \code{cluster_manual_causa}: ...
\item \code{causa}: ...
\item \code{clean_causa}: ...
\item \code{consecuencia}: ...
\item \code{clean_consecuencia}: ...
\item \code{region}: ...
}
}
\source{
Los datos fueron recopilados del laboratorio de instrumentos.
}
\usage{
labinstrumentos
}
\description{
Datos, análisis y visualizaciones del laboratorio de intrumentos.
}
\keyword{datasets}
