% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_datos.R
\name{obtener_datos}
\alias{obtener_datos}
\title{Obtener los datos actualizados y limpios}
\usage{
obtener_datos()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Llamar a los datos alojados en el Google sheets con los datos crudos
}
