% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_frecuencias.R
\name{obtener_frecuencias}
\alias{obtener_frecuencias}
\title{Obtener frecuencia de palabras}
\usage{
obtener_frecuencias(i, n_words = 1)
}
\arguments{
\item{i}{vector de carácteres.}

\item{n_words}{Número de palabras combinadas.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
