% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_sankey.R
\name{crear_sankey}
\alias{crear_sankey}
\title{Crea un gráfico sankey con uno o más términos}
\usage{
crear_sankey(
  datos,
  n_words = 1,
  freq_minima = 3,
  vars = c("clean_problema", "clean_causa", "clean_consecuencia"),
  height = 1000,
  output = "grafico"
)
}
\arguments{
\item{datos}{Datos usados para la generación del gráfico.}

\item{n_words}{Número de palabras combinadas. Por defecto es 1.}

\item{freq_minima}{Frecuencia mínima de palabras para filtrar. Por defecto es 3.}

\item{vars}{Variables para visualizar. Por defecto son las variables \code{clean_problema}, \code{clean_causa} y \code{clean_consecuencia}.}

\item{height}{Altura en píxeles del gråfico.}

\item{output}{Carácter de longitud uno. Tipo de objeto resultante. Las opciones
son \code{"grafico"} (por defecto), \code{"df_largo"}, \code{"df_corto"}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Apoyándose del paquete \code{highcharter}, crea un diagrama Sankey para visualizar
la relación entre términos de palabras
}
