% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labinstrumentos-package.R
\docType{package}
\name{labinstrumentos-package}
\alias{labinstrumentos}
\alias{labinstrumentos-package}
\title{Datos del Laboratorio de Instrumentos}
\description{
Datos, análisis y visualizaciones del laboratorio de intrumentos.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SEREMICTCI/labinstrumentos}
  \item Report bugs at \url{https://github.com/SEREMICTCI/labinstrumentos/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Carlos Morales Quiroz \email{csmorales@minciencia.gob.cl} [contributor]
}

}
\keyword{internal}
